+++ common/src/main/kotlin/com/cobblemon/mod/common/api/molang/MoLangFunctions.kt
@@ -40,6 +40,7 @@ import com.cobblemon.mod.common.api.moves.animations.ActionEffectContext
 import com.cobblemon.mod.common.api.moves.animations.ActionEffects
 import com.cobblemon.mod.common.api.moves.animations.NPCProvider
 import com.cobblemon.mod.common.api.moves.animations.TargetsProvider
+import com.cobblemon.mod.common.api.npc.NPCClass
 import com.cobblemon.mod.common.api.npc.NPCClasses
 import com.cobblemon.mod.common.api.npc.configuration.interaction.DialogueNPCInteractionConfiguration
 import com.cobblemon.mod.common.api.npc.configuration.interaction.ScriptNPCInteractionConfiguration
@@ -56,6 +57,7 @@ import com.cobblemon.mod.common.api.pokemon.evolution.Evolution
 import com.cobblemon.mod.common.api.pokemon.experience.SidemodExperienceSource
 import com.cobblemon.mod.common.api.pokemon.moves.LearnsetQuery
 import com.cobblemon.mod.common.api.pokemon.stats.Stats
+import com.cobblemon.mod.common.api.riding.Rideable
 import com.cobblemon.mod.common.api.scheduling.ClientTaskTracker
 import com.cobblemon.mod.common.api.scheduling.Schedulable
 import com.cobblemon.mod.common.api.scheduling.ServerTaskTracker
@@ -84,6 +86,7 @@ import com.cobblemon.mod.common.entity.pokemon.PokemonBehaviourFlag
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity
 import com.cobblemon.mod.common.entity.pokemon.ai.PokemonMoveControl
 import com.cobblemon.mod.common.net.messages.client.animation.PlayPosableAnimationPacket
+import com.cobblemon.mod.common.net.messages.client.battle.BattleMusicPacket
 import com.cobblemon.mod.common.net.messages.client.effect.RunPosableMoLangPacket
 import com.cobblemon.mod.common.net.messages.client.effect.SpawnSnowstormEntityParticlePacket
 import com.cobblemon.mod.common.net.messages.client.effect.SpawnSnowstormParticlePacket
@@ -125,21 +128,28 @@ import net.minecraft.sounds.SoundSource
 import net.minecraft.tags.TagKey
 import net.minecraft.world.damagesource.DamageSource
 import net.minecraft.world.damagesource.DamageTypes
+import net.minecraft.world.effect.MobEffectInstance
 import net.minecraft.world.entity.Entity
 import net.minecraft.world.entity.EntityDimensions
 import net.minecraft.world.entity.EntityType
 import net.minecraft.world.entity.LightningBolt
 import net.minecraft.world.entity.LivingEntity
+import net.minecraft.world.entity.Mob
 import net.minecraft.world.entity.PathfinderMob
+import net.minecraft.world.entity.PlayerRideable
 import net.minecraft.world.entity.TamableAnimal
 import net.minecraft.world.entity.ai.behavior.BlockPosTracker
 import net.minecraft.world.entity.ai.memory.MemoryModuleType
 import net.minecraft.world.entity.ai.memory.MemoryStatus
 import net.minecraft.world.entity.ai.memory.WalkTarget
+import net.minecraft.world.entity.animal.Animal
 import net.minecraft.world.entity.item.ItemEntity
+import net.minecraft.world.entity.monster.Enemy
+import net.minecraft.world.entity.monster.Monster
 import net.minecraft.world.entity.player.Player
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.level.ClipContext
+import net.minecraft.world.level.GameType
 import net.minecraft.world.level.Level
 import net.minecraft.world.level.Level.ExplosionInteraction
 import net.minecraft.world.level.biome.Biome
@@ -149,6 +159,9 @@ import net.minecraft.world.level.levelgen.Heightmap
 import net.minecraft.world.level.pathfinder.PathType
 import net.minecraft.world.phys.AABB
 import net.minecraft.world.phys.Vec3
+import kotlin.text.get
+import kotlin.text.toInt
+import kotlin.toString
 
 /**
  * Holds a bunch of useful MoLang trickery that can be used or extended in API
@@ -198,6 +211,14 @@ object MoLangFunctions {
         },
         "string_length" to java.util.function.Function { params ->
             return@Function DoubleValue(params.getString(0).length)
+        },
+        "split_string" to java.util.function.Function { params ->
+            val text = params.getString(0)
+            val delimiter = params.getString(1)
+            val parts = text.split(delimiter).map { StringValue(it) }
+            val struct = ArrayStruct(hashMapOf())
+            parts.forEachIndexed { index, moValue -> struct.setDirectly("$index", moValue) }
+            return@Function struct
         },
         "is_blank" to java.util.function.Function { params ->
             val arg = params.get<MoValue>(0)
@@ -466,6 +487,29 @@ object MoLangFunctions {
                     return@put DoubleValue.ZERO
                 }
             }
+            map.put("spawn_npc") { params ->
+                val x = params.getInt(0)
+                val y = params.getInt(1)
+                val z = params.getInt(2)
+                val npcClass = params.getString(3)
+
+                val pos = BlockPos(x, y, z)
+
+                if (!Level.isInSpawnableBounds(pos)) {
+                    return@put DoubleValue.ZERO
+                }
+
+                val npc = NPCEntity(world) ?: return@put DoubleValue.ZERO
+                npc.moveTo(pos, npc.yRot, npc.xRot)
+                npc.npc = NPCClasses.getByIdentifier(npcClass.asIdentifierDefaultingNamespace()) ?: return@put DoubleValue.ZERO
+                npc.initialize(1)
+
+                if (world.addFreshEntity(npc)) {
+                    return@put npc.struct
+                } else {
+                    return@put DoubleValue.ZERO
+                }
+            }
             map.put("play_sound_on_server") { params ->
                 val sound = params.getString(0).asResource()
                 val soundSource = params.getString(1).uppercase()
@@ -550,10 +594,61 @@ object MoLangFunctions {
                 environment
             }
             map.put("is_player") { DoubleValue.ONE }
+            map.put("riding_pokemon") {
+                val vehicle = player.vehicle
+                if (vehicle is PokemonEntity) {
+                    return@put vehicle.struct
+                } else {
+                    return@put DoubleValue.ZERO
+                }
+            }
             if (player is ServerPlayer) {
+                map.put("seen_credits") { _ ->
+                    DoubleValue( DoubleValue(player.seenCredits) )
+                }
+                map.put("is_in_dialogue") { _ ->
+                    DoubleValue( DoubleValue(player.isInDialogue) )
+                }
+                map.put("active_dialogue") { _ ->
+                    if (player.isInDialogue) {
+                        player.activeDialogue?.dialogueId.toString()
+                        return@put DoubleValue.ONE
+                    } else {
+                        DoubleValue.ZERO
+                    }
+                }
+                map.put("is_spectator") { DoubleValue(player.isSpectator) }
+                map.put("is_creative") { DoubleValue(player.isCreative) }
+                map.put("is_survival") { DoubleValue(player.gameMode.isSurvival) }
+                map.put("is_adventure") { DoubleValue(player.gameMode.getGameModeForPlayer() == GameType.ADVENTURE) }
                 map.put("run_command") { params ->
                     val command = params.getString(0)
                     player.server.commands.performPrefixedCommand(player.createCommandSourceStack(), command)
+                }
+                map.put("set_battle_theme") { params ->
+                    val soundId = params.getString(0).asResource()
+                    Cobblemon.playerDataManager.getGenericData(player).battleTheme = soundId
+                    return@put DoubleValue.ONE
+                }
+                map.put("battle_music") { params ->
+                    val soundId = params.getString(0).asResource()
+                    val volume = params.getDoubleOrNull(1)?.toFloat() ?: 1.0f
+                    val pitch = params.getDoubleOrNull(2)?.toFloat() ?: 1.0f
+                    val restart = params.getBooleanOrNull(3) ?: true
+
+                    val music = soundId
+                    if (music != null) {
+                        val packet = BattleMusicPacket(music, volume, pitch, restart)
+                        packet.sendToPlayer(player)
+                        return@put DoubleValue.ONE
+                    } else {
+                        return@put DoubleValue.ZERO
+                    }
+                }
+                map.put("stop_battle_music") { _ ->
+                    val packet = BattleMusicPacket(null)
+                    packet.sendToPlayer(player)
+                    return@put DoubleValue.ONE
                 }
                 map.put("play_sound_on_server") { params ->
                     val sound = params.getString(0).asResource()
@@ -734,7 +829,26 @@ object MoLangFunctions {
             map.put("delta_movement") {
                 return@put listOf(entity.deltaMovement.x, entity.deltaMovement.y, entity.deltaMovement.z).asArrayValue(::DoubleValue)
             }
-
+            map.put("tags") {
+                val tags = entity.tags
+                val array = ArrayStruct(hashMapOf())
+                tags.forEachIndexed { index, tag -> array.setDirectly("$index", StringValue(tag)) }
+                return@put array
+            }
+            map.put("add_tag") { params ->
+                val tag = params.getString(0)
+                entity.addTag(tag)
+                return@put DoubleValue.ONE
+            }
+            map.put("remove_tag") { params ->
+                val tag = params.getString(0)
+                entity.removeTag(tag)
+                return@put DoubleValue.ONE
+            }
+            map.put("has_tag") { params ->
+                val tag = params.getString(0)
+                return@put DoubleValue(entity.tags.contains(tag))
+            }
             if (entity is MoLangScriptingEntity) {
                 map.put("add_callback") { params ->
                     val type = params.getString(0).asIdentifierDefaultingNamespace()
@@ -942,6 +1056,11 @@ object MoLangFunctions {
                     packet.sendToPlayer(player)
                 }
             }
+            map.put("make_intangible") { params ->
+                val intangible = params.getBooleanOrNull(0) ?: true
+                entity.noPhysics = intangible
+                return@put DoubleValue.ONE
+            }
             return@mutableListOf map
         }
     )
@@ -949,9 +1068,56 @@ object MoLangFunctions {
     val livingEntityFunctions: MutableList<(LivingEntity) -> HashMap<String, java.util.function.Function<MoParams, Any>>> = mutableListOf<(LivingEntity) -> HashMap<String, java.util.function.Function<MoParams, Any>>>(
         { entity ->
             val map = hashMapOf<String, java.util.function.Function<MoParams, Any>>()
+            map.put("is_player") { _ -> DoubleValue(entity is Player) }
+            map.put("is_npc") { _ -> DoubleValue(entity is NPCEntity) }
+            map.put("is_mob") { _ -> DoubleValue(entity is Mob) }
+            map.put("is_pokemon") { _ -> DoubleValue(entity is PokemonEntity) }
+            map.put("is_animal") { _ -> DoubleValue(entity is Animal) }
+            map.put("is_tamable") { _ -> DoubleValue(entity is TamableAnimal) }
+            map.put("is_tamed") { _ -> DoubleValue(entity is TamableAnimal && entity.isTame) }
+            map.put("is_hostile") { _ -> DoubleValue(entity is Monster) }
+            map.put("is_baby") { _ -> DoubleValue(entity.isBaby) }
+            map.put("is_adult") { _ -> DoubleValue(!entity.isBaby) }
+            map.put("remove_effect") { params ->
+                val effectId = params.getString(0).asIdentifierDefaultingNamespace()
+                val effectHolder = BuiltInRegistries.MOB_EFFECT.getHolder(effectId).orElse(null)
+                if (effectHolder != null) {
+                    entity.removeEffect(effectHolder)
+                    return@put DoubleValue.ONE
+                }
+            }
+            map.put("add_effect") { params ->
+                val effectId = params.getString(0).asIdentifierDefaultingNamespace()
+                val duration = params.getInt(1)
+                val amplifier = params.getIntOrNull(2) ?: 0
+                val ambient = params.getBooleanOrNull(3) ?: false
+                val visible = params.getBooleanOrNull(4) ?: true
+                val effectHolder = BuiltInRegistries.MOB_EFFECT.getHolder(effectId).orElse(null)
+                if (effectHolder != null) {
+                    entity.addEffect(MobEffectInstance(effectHolder, duration, amplifier, ambient, visible))
+                    return@put DoubleValue.ONE
+                }
+            }
+            map.put("has_effect") { params ->
+                val effectId = params.getString(0).asIdentifierDefaultingNamespace()
+                val effectHolder = BuiltInRegistries.MOB_EFFECT.getHolder(effectId).orElse(null)
+                return@put DoubleValue(if (effectHolder != null) entity.hasEffect(effectHolder) else false)
+            }
             map.put("heal") { params ->
                 val amount = params.getDouble(0)
                 entity.heal(amount.toFloat())
+            }
+            map.put("is_looking_at") { params ->
+                val targetEntity = params.get<MoValue>(0)
+                val maxDistance = params.getDoubleOrNull(1)?.toFloat() ?: 2.0F
+
+                val entity = if (targetEntity is ObjectValue<*> && targetEntity.obj is Entity) {
+                    targetEntity.obj as Entity
+                } else {
+                    return@put DoubleValue.ZERO
+                }
+
+                return@put DoubleValue(entity.isLookingAt(entity, maxDistance))
             }
             map.put("is_living_entity") { DoubleValue.ONE }
             map.put("is_flying") { _ -> DoubleValue(entity.isFallFlying) }
@@ -1075,7 +1241,6 @@ object MoLangFunctions {
                     DoubleValue(entity.brain.getMemory(MemoryModuleType.WALK_TARGET).isPresent || entity.isPathFinding)
                 }
             }
-
             map
         }
     )
@@ -1296,6 +1461,11 @@ object MoLangFunctions {
             map.put("has_party") { DoubleValue(npc.party != null) }
             map.put("is_npc") { DoubleValue.ONE }
             map.put("can_battle") { DoubleValue(npc.party?.any { it.currentHealth > 0 } == true || npc.npc.party?.isStatic == false) }
+            map.put("set_battle_theme") { params ->
+                val soundId = params.getString(0).asResource()
+                npc.npc.battleTheme = soundId
+                return@put DoubleValue.ONE
+            }
             map
         }
     )
@@ -1419,7 +1589,7 @@ object MoLangFunctions {
                 DoubleValue(pokemon.species.nationalPokedexNumber.toDouble())
             }
             map.put("types") {
-               pokemon.form.types.map { it.toString() }.asArrayValue(::StringValue)
+                pokemon.form.types.map { it.toString() }.asArrayValue(::StringValue)
             }
             map.put("gender_ratio") {
                 DoubleValue(pokemon.form.maleRatio.toDouble())
@@ -1664,7 +1834,11 @@ object MoLangFunctions {
                 val moveTemplate = Moves.getByName(moveName) ?: return@put DoubleValue.ZERO
                 val includeLegacy = params.getBooleanOrNull(1) ?: true
 
-                val canLearn = LearnsetQuery.ANY.canLearn(moveTemplate, pokemon.form.moves)
+                val canLearn = if (includeLegacy) {
+                    LearnsetQuery.ANY.canLearn(moveTemplate, pokemon.form.moves)
+                } else {
+                    LearnsetQuery.LEGAL.canLearn(moveTemplate, pokemon.form.moves)
+                }
                 return@put DoubleValue(if (canLearn) 1.0 else 0.0)
             }
             map.put("unlearn_move") { params ->
@@ -1677,6 +1851,15 @@ object MoLangFunctions {
                     return@put DoubleValue.ZERO
                 }
             }
+            map.put("can_evolve") { _ ->
+                DoubleValue(pokemon.evolutions.any())
+            }
+            map.put("force_evolve") { params ->
+                val idx = params.getInt(0)
+                val evolution = if (idx >= 0) pokemon.evolutions.elementAtOrNull(idx) else return@put DoubleValue.ZERO
+                evolution?.forceEvolve(pokemon)
+                StringValue(evolution.toString())
+            }
             map
         }
     )
@@ -1684,6 +1867,7 @@ object MoLangFunctions {
     val pokemonEntityFunctions = mutableListOf<(PokemonEntity) -> HashMap<String, java.util.function.Function<MoParams, Any>>>(
         { pokemonEntity ->
             val map = hashMapOf<String, java.util.function.Function<MoParams, Any>>()
+            map.put("is_busy") { DoubleValue(pokemonEntity.isBusy) }
             map.put("in_battle") { DoubleValue(pokemonEntity.isBattling) }
             map.put("is_moving") { DoubleValue((pokemonEntity.moveControl as? PokemonMoveControl)?.hasWanted() == true) }
             map.put("is_flying") { DoubleValue(pokemonEntity.getBehaviourFlag(PokemonBehaviourFlag.FLYING)) }
@@ -2449,3 +2633,4 @@ object MoLangFunctions {
 }
 
 fun Either<ResourceLocation, ExpressionLike>.runScript(runtime: MoLangRuntime) = map({ CobblemonScripts.run(it, runtime) }, { it.resolve(runtime) })
+